+++ sdk/src/main/java/com/hedera/hashgraph/sdk/AccountAllowanceDeleteTransaction.java
@@ -41,12 +41,6 @@ public class AccountAllowanceDeleteTransaction extends com.hedera.hashgraph.sdk.
 
     private void initFromTransactionBody() {
         var body = sourceTransactionBody.getCryptoDeleteAllowance();
-        for (var allowanceProto : body.getCryptoAllowancesList()) {
-            hbarAllowances.add(HbarAllowance.fromProtobuf(allowanceProto));
-        }
-        for (var allowanceProto : body.getTokenAllowancesList()) {
-            tokenAllowances.add(TokenAllowance.fromProtobuf(allowanceProto));
-        }
         for (var allowanceProto : body.getNftAllowancesList()) {
             getNftSerials(
                 AccountId.fromProtobuf(allowanceProto.getOwner()),
@@ -55,16 +49,28 @@ public class AccountAllowanceDeleteTransaction extends com.hedera.hashgraph.sdk.
         }
     }
 
+    /**
+     * @deprecated with no replacement
+     */
+    @Deprecated
     public AccountAllowanceDeleteTransaction deleteAllHbarAllowances(AccountId ownerAccountId) {
         requireNotFrozen();
         hbarAllowances.add(new HbarAllowance(Objects.requireNonNull(ownerAccountId), null, null));
         return this;
     }
 
+    /**
+     * @deprecated with no replacement
+     */
+    @Deprecated
     public List<HbarAllowance> getHbarAllowanceDeletions() {
         return new ArrayList<>(hbarAllowances);
     }
 
+    /**
+     * @deprecated with no replacement
+     */
+    @Deprecated
     public AccountAllowanceDeleteTransaction deleteAllTokenAllowances(TokenId tokenId, AccountId ownerAccountId) {
         requireNotFrozen();
         tokenAllowances.add(new TokenAllowance(
@@ -76,6 +82,10 @@ public class AccountAllowanceDeleteTransaction extends com.hedera.hashgraph.sdk.
         return this;
     }
 
+    /**
+     * @deprecated with no replacement
+     */
+    @Deprecated
     public List<TokenAllowance> getTokenAllowanceDeletions() {
         return new ArrayList<>(tokenAllowances);
     }
@@ -135,12 +145,6 @@ public class AccountAllowanceDeleteTransaction extends com.hedera.hashgraph.sdk.
 
     CryptoDeleteAllowanceTransactionBody.Builder build() {
         var builder = CryptoDeleteAllowanceTransactionBody.newBuilder();
-        for (var allowance : hbarAllowances) {
-            builder.addCryptoAllowances(allowance.toWipeProtobuf());
-        }
-        for (var allowance : tokenAllowances) {
-            builder.addTokenAllowances(allowance.toWipeProtobuf());
-        }
         for (var allowance : nftAllowances) {
             builder.addNftAllowances(allowance.toRemoveProtobuf());
         }
@@ -159,12 +163,6 @@ public class AccountAllowanceDeleteTransaction extends com.hedera.hashgraph.sdk.
 
     @Override
     void validateChecksums(Client client) throws BadEntityIdException {
-        for (var allowance : hbarAllowances) {
-            allowance.validateChecksums(client);
-        }
-        for (var allowance : tokenAllowances) {
-            allowance.validateChecksums(client);
-        }
         for (var allowance : nftAllowances) {
             allowance.validateChecksums(client);
         }