+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Deprecated
+
+ * `AccountAllowanceAdjustTransaction` with no replacement.
+ * `AccountAllowanceDeleteTransaction.deleteAll[Hbar|Token]Allowances` with no replacements.
+ * `TransactionRecord.[hbar|token|tokenNft]AllowanceAdjustments`.
+ * `AccountInfo.[hbar|token|tokenNft]Allowances`.
+
 ## v2.12.0-beta.1
 
 ### Added